+++ src/XmlScope.spec.ts
@@ -75,8 +75,7 @@ describe('XmlScope', () => {
                 <component name="ChildComponent" extends="ParentComponent">
                 </component>
             `);
-            let childScope = program.getScopesForFile(childXmlFile);
-            let definition = childScope[0].getDefinition(childXmlFile, Position.create(1, 48));
+            const definition = program.getDefinition(childXmlFile.srcPath, Position.create(1, 48));
             expect(definition).to.be.lengthOf(1);
             expect(definition[0].uri).to.equal(util.pathToUri(parentXmlFile.srcPath));
         });