+++ src/bscPlugin/BscPlugin.ts
@@ -1,10 +1,8 @@
 import { isBrsFile, isXmlFile } from '../astUtils/reflection';
-import type { CompilerPlugin, OnFileValidateEvent, OnGetCodeActionsEvent, ProvideHoverEvent, OnGetSemanticTokensEvent, OnScopeValidateEvent, ProvideCompletionsEvent, AfterProgramValidateEvent, ProvideFileEvent, AfterSerializeFileEvent, BeforeBuildProgramEvent, WriteFileEvent, OnPrepareFileEvent } from '../interfaces';
+import type { CompilerPlugin, OnFileValidateEvent, OnGetCodeActionsEvent, ProvideHoverEvent, OnGetSemanticTokensEvent, OnScopeValidateEvent, ProvideCompletionsEvent, ProvideDefinitionEvent, AfterProgramValidateEvent, AfterSerializeFileEvent, BeforeBuildProgramEvent, OnPrepareFileEvent, ProvideFileEvent, WriteFileEvent } from '../interfaces';
 import { CodeActionsProcessor } from './codeActions/CodeActionsProcessor';
 import { CompletionsProcessor } from './completions/CompletionsProcessor';
-import { FileProvider } from './fileProviders/FileProvider';
-import { FileSerializer } from './serialize/FileSerializer';
-import { FileWriter } from './FileWriter';
+import { DefinitionProvider } from './definition/DefinitionProvider';
 import { HoverProcessor } from './hover/HoverProcessor';
 import { BrsFileSemanticTokensProcessor } from './semanticTokens/BrsFileSemanticTokensProcessor';
 import { BrsFileValidator } from './validation/BrsFileValidator';
@@ -16,6 +14,9 @@ import { BrsFilePreTranspileProcessor } from './transpile/BrsFileTranspileProces
 import { XmlFilePreTranspileProcessor } from './transpile/XmlFilePreTranspileProcessor';
 import type { BrsFile } from '../files/BrsFile';
 import type { XmlFile } from '../files/XmlFile';
+import { FileWriter } from './FileWriter';
+import { FileProvider } from './fileProviders/FileProvider';
+import { FileSerializer } from './serialize/FileSerializer';
 
 export class BscPlugin implements CompilerPlugin {
     public name = 'BscPlugin';
@@ -34,6 +35,10 @@ export class BscPlugin implements CompilerPlugin {
 
     public provideCompletions(event: ProvideCompletionsEvent) {
         new CompletionsProcessor(event).process();
+    }
+
+    public provideDefinition(event: ProvideDefinitionEvent) {
+        new DefinitionProvider(event).process();
     }
 
     public onGetSemanticTokens(event: OnGetSemanticTokensEvent) {